# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexusGraph - Federated GraphQL SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List all organizations
      parameters:
        - name: plan
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationList"

    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get an organization by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationBody"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"

    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Billing
...
  /billing/subscriptions/{orgId}:
    get:
      operationId: getSubscription
      tags: [billing]
      summary: Get subscription for an organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

    patch:
      operationId: updateSubscription
      tags: [billing]
      summary: Update subscription plan
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionBody"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # Subgraphs
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, developer, viewer]
        status:
          type: string
          enum: [active, suspended, pending]
        orgId:
          type: ["integer", "null"]
        orgName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        lastActiveAt:
          type: ["string", "null"]
      required:
        - id
        - email
        - name
        - role
        - status
        - createdAt
...
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        domain:
          type: ["string", "null"]
        planId:
          type: integer
        planName:
          type: string
        status:
          type: string
          enum: [active, suspended, trialing, cancelled]
        userCount:
          type: integer
        subgraphCount:
          type: integer
        monthlyOperations:
          type: integer
        createdAt:
          type: string
        logoUrl:
          type: ["string", "null"]
      required:
        - id
        - name
        - slug
        - planId
        - planName
        - status
        - userCount
        - subgraphCount
        - monthlyOperations
        - createdAt
    OrganizationList:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        total:
          type: integer
      required:
        - organizations
        - total
    CreateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        domain:
          type: ["string", "null"]
        planId:
          type: integer
      required:
        - name
        - slug
        - planId
    UpdateOrganizationBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        domain:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [active, suspended, trialing, cancelled, null]
    BillingPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        priceMonthly:
          type: number
        priceYearly:
          type: number
        maxUsers:
          type: ["integer", "null"]
        maxSubgraphs:
          type: ["integer", "null"]
        maxMonthlyOperations:
          type: ["integer", "null"]
        features:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - slug
        - priceMonthly
        - priceYearly
        - features
    Subscription:
      type: object
      properties:
        orgId:
          type: integer
        orgName:
          type: string
        plan:
          $ref: "#/components/schemas/BillingPlan"
        status:
          type: string
          enum: [active, trialing, cancelled, past_due]
        currentPeriodStart:
          type: string
        currentPeriodEnd:
          type: string
        cancelAtPeriodEnd:
          type: boolean
      required:
        - orgId
        - orgName
        - plan
        - status
        - currentPeriodStart
        - currentPeriodEnd
        - cancelAtPeriodEnd
    UpdateSubscriptionBody:
      type: object
      properties:
        planId:
          type: integer
        cancelAtPeriodEnd:
          type: ["boolean", "null"]
      required:
        - planId
...