# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexusGraph - Federated GraphQL SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  orgId?: number | null;
  /** @nullable */
  orgName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
  /** @nullable */
  lastActiveAt?: string | null;
}